- LastFmSqliteBackend: ADO.NET is again leading (giving up on entity framework). - LastFmSqliteBackend: removed triggers.  Implemented use of Current...List in list setters and getters. - PlaylistFixer/Program.cs: updated to deal with playlists that contain file names only. - PlaylistSuggestions: completely rewritten algorithm that should better be able to capitalize on a large local cache.  seems to work great ;-) - SongDiskCrawler/SongDiskCrawlerMain.cs temporary extra reading to debug taglibsharp - taglibsharp (not in this repos): fixed unsynchronization bug (see http://bugzilla.gnome.org/show_bug.cgi?id=593138) - removed file.OpenWrite calls which can leave behind old junk. - Several minor improvements and refactorings.

+++ SongSearchLinq/LastFMspider/LastFMSQLiteBackend/InsertArtistSimilarity.cs
@@ -30,7 +30,7 @@ WHERE B.LowercaseArtist= @lowerArtistB
       
 
 
-        public void Execute(int listID, string artistB, double rating) {
+        public void Execute(long listID, string artistB, double rating) {
             lock (SyncRoot) {
                 lfmCache.InsertArtist.Execute(artistB);//we could also replace casing... whatever...
